// src/app/interfaces/applicant-registration.ts
export interface ApplicantRegistration {

  username: string |null | undefined;
  password: string |null | undefined ; 
  email: string |null | undefined ; 
  Applicant_firstname: string |null | undefined ; 
  Applicant_lastname: string |null | undefined ; 
  user_phone_number: string |null | undefined ; 
  user_join_date: string |null | undefined ; 
  Applicant_birthday: string |null | undefined ; 
  notifications: any[] |null | undefined ; 
  alerts: any[] |null | undefined ; 
  applicantProfile: any |null | undefined ; 
}